SmallFix

+++ models/DailyMenuOrder.go
@@ -11,7 +11,7 @@ type DailyMenuOrder struct {
 	bun.BaseModel `bun:"daily_menu_orders,select:daily_menu_orders_view,alias:daily_menu_orders_view"`
 	ID            uuid.NullUUID `bun:"id,pk,type:uuid,default:uuid_generate_v4()" json:"id" `
 	Date          time.Time     `bun:"item_date" json:"date"`
-	BoxNumber     uint8         `json:"boxNumber"`
+	BoxNumber     uint          `json:"boxNumber"`
 	Number        uint          `bun:",notnull,autoincrement" json:"number"`
 
 	FormValue   *FormValue    `bun:"rel:belongs-to" json:"formValue"`
