+++ models/DailyMenuOrderItem.go
@@ -9,8 +9,8 @@ type DailyMenuOrderItem struct {
 	bun.BaseModel `bun:"daily_menu_order_items,alias:daily_menu_order_items"`
 	ID            uuid.NullUUID `bun:"id,pk,type:uuid,default:uuid_generate_v4()" json:"id" `
 
-	Quantity uint8 `json:"quantity"`
-	Price    uint8 `json:"price"`
+	Quantity uint `json:"quantity"`
+	Price    uint `json:"price"`
 
 	DailyMenuOrder   *DailyMenuOrder `bun:"rel:belongs-to" json:"dailyMenuOrder"`
 	DailyMenuOrderID uuid.NullUUID   `bun:"type:uuid"  json:"dailyMenuOrderId"`
